+++ lib/Remote/RemoteWebDriver.php
@@ -15,8 +15,10 @@
 
 namespace Facebook\WebDriver\Remote;
 
+use Facebook\WebDriver\Html5\WebStorage;
 use Facebook\WebDriver\Interactions\WebDriverActions;
 use Facebook\WebDriver\JavaScriptExecutor;
+use Facebook\WebDriver\Remote\Html5\RemoteWebStorage;
 use Facebook\WebDriver\WebDriver;
 use Facebook\WebDriver\WebDriverBy;
 use Facebook\WebDriver\WebDriverCapabilities;
@@ -404,6 +406,16 @@ class RemoteWebDriver implements WebDriver, JavaScriptExecutor, WebDriverHasInpu
     public function switchTo()
     {
         return new RemoteTargetLocator($this->getExecuteMethod(), $this);
+    }
+
+    /**
+     * An abstraction allowing the driver to access the browser's storage.
+     *
+     * @return WebStorage
+     */
+    public function storage()
+    {
+        return new RemoteWebStorage($this->getExecuteMethod());
     }
 
     /**