+++ lib/WebDriver.php
@@ -15,6 +15,7 @@
 
 namespace Facebook\WebDriver;
 
+use Facebook\WebDriver\Html5\WebStorage;
 use Facebook\WebDriver\Interactions\Touch\WebDriverTouchScreen;
 
 /**
@@ -110,6 +111,14 @@ interface WebDriver extends WebDriverSearchContext
      * @return WebDriverOptions
      */
     public function manage();
+
+    // TODO: Add in next major release (BC)
+    ///**
+    // * An abstraction allowing the driver to access the browser's storage.
+    // *
+    // * @return WebStorage
+    // */
+    //public function storage(): WebStorage;
 
     /**
      * An abstraction allowing the driver to access the browser's history and to